chore: relocating base objects, change component name
