--- src/main/kotlin/com/picpin/api/aspect/AccountId.kt
@@ -1,5 +1,0 @@
-package com.picpin.api.aspect
-
-@Target(AnnotationTarget.VALUE_PARAMETER)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class AccountId()
