+++ src/main/kotlin/com/picpin/api/domains/account/AccountRepository.kt
@@ -1,7 +1,7 @@
 package com.picpin.api.domains.account
 
-import com.picpin.api.verticals.domain.BusinessErrorCode
-import com.picpin.api.verticals.domain.BusinessException
+import com.picpin.api.verticals.domain.exception.BusinessErrorCode
+import com.picpin.api.verticals.domain.exception.BusinessException
 import jakarta.persistence.LockModeType
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Lock
@@ -14,7 +14,7 @@ interface AccountRepository : JpaRepository<Account, Long> {
     fun findByVendorId(vendorId: Long): Account
 }
 
-fun AccountRepository.findBy(accountId: Long): Account {
+fun AccountRepository.findOneOrThrow(accountId: Long): Account {
     return findById(accountId).orElseThrow {
         BusinessException.of(BusinessErrorCode.ACCOUNT_NOT_FOUND, "accountId = $accountId")
     }
