+++ src/main/kotlin/com/picpin/api/domains/oauth/AccessTokenParser.kt
@@ -3,8 +3,8 @@ package com.picpin.api.domains.oauth
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
 import com.auth0.jwt.interfaces.DecodedJWT
-import com.picpin.api.verticals.domain.BusinessErrorCode
-import com.picpin.api.verticals.domain.BusinessException
+import com.picpin.api.verticals.domain.exception.BusinessErrorCode
+import com.picpin.api.verticals.domain.exception.BusinessException
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.nio.charset.StandardCharsets
@@ -24,7 +24,7 @@ class AccessTokenParser(
                 .verify(accessToken)
         } catch (exception: Exception) {
             throw BusinessException.of(
-                BusinessErrorCode.INVALID_JWT_TOKEN,
+                BusinessErrorCode.INVALID_VERIFIED_JWT_TOKEN,
                 "${exception.localizedMessage}, accessToken = $accessToken"
             )
         }
