+++ src/main/kotlin/com/picpin/api/interfaces/MetaApi.kt
@@ -1,9 +1,9 @@
 package com.picpin.api.interfaces
 
-import com.picpin.api.domains.base.staticProvinceFilters
-import com.picpin.api.domains.base.staticMarkerColors
-import com.picpin.api.domains.base.staticMarkerColorFilters
+import com.picpin.api.interfaces.model.staticProvinceFilters
 import com.picpin.api.interfaces.model.GetMetaAttributesResponse
+import com.picpin.api.interfaces.model.staticMarkerColorFilters
+import com.picpin.api.interfaces.model.staticMarkerColors
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
