--- src/main/kotlin/com/picpin/api/verticals/web/exception/ExceptionResponse.java
@@ -1,7 +1,0 @@
-package com.picpin.api.verticals.web.exception;
-
-public record ExceptionResponse(
-    int errorCode
-) {
-
-}
